Relocation of the test, and removing classes that are no longer used
