--- apex-squid/src/main/java/org/fundacionjala/enforce/sonarqube/apex/api/grammar/head/ApexGrammarModifier.java
@@ -1,52 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2016 Jalasoft.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.fundacionjala.enforce.sonarqube.apex.api.grammar.head;
-
-import static org.fundacionjala.enforce.sonarqube.apex.api.grammar.RuleKey.KEYWORD;
-import static org.fundacionjala.enforce.sonarqube.apex.api.grammar.RuleKey.LOOKAHEAD;
-import static org.fundacionjala.enforce.sonarqube.apex.api.grammar.RuleKey.LOOKAHEAD_KEYWORD;
-import static org.fundacionjala.enforce.sonarqube.apex.api.grammar.RuleKey.MODIFIER;
-import org.sonar.sslr.grammar.LexerlessGrammarBuilder;
-
-/**
- * The class Lookahead and Keyword joins rules, and create rule Modifier.
- */
-public class ApexGrammarModifier {
-
-    /**
-     * Grammar is created for a class modifier (lookahead or keyword).
-     *
-     * @return The grammar for the modifier.
-     */
-    public static LexerlessGrammarBuilder createGrammarBuilder() {
-        LexerlessGrammarBuilder grammarBuilder = LexerlessGrammarBuilder.create();
-        grammarBuilder.rule(LOOKAHEAD_KEYWORD).is(ApexGrammarLookahead.createGrammarBuilder().build().rule(LOOKAHEAD),
-                ApexGrammarKeyword.createGrammarBuilder().build().rule(KEYWORD));
-        grammarBuilder.rule(MODIFIER).is(grammarBuilder.firstOf(
-                LOOKAHEAD_KEYWORD,
-                ApexGrammarLookahead.createGrammarBuilder().build().rule(LOOKAHEAD))
-        );
-        return grammarBuilder;
-    }
-}
